Fixed: #135 Empty password reset form submits as if valid #133 Settings form to reset password should give more descriptive errors
#135:
When allowing users to update their settings, an issue occurred where any time a password is not included as part of a user model update change the form breaks. To correct this I removed the must exist validation for password forms. The consequence of this is the user could successfully submit a password reset request with a blank password (which would not actually change their password but look like it did). I created a new update_password controller to handle the case of updating passwords and independently verify the password exists. All other validations work. This controller takes place of the update_attributes controller in the case of updating passwords.
Users can still generate a users#update request that specifies a blank password that does not return an error, but only through API and the request won't actually succeed (just look like it does). Ideally there's a better way to fix this but I can't find it.

#133:
Rather than re-rendering the form (redirect_to :back), which loses the session data as far as I know. I specifically re-render the form and pass the user errors. Also removed the existing flash in this case.
Cleaned up some messiness in the user model that tried to maintain the not-logged-in session info on this form and simply sign in the user before rendering the form. No advantage to doing it the way I had and it added unecessary complexity.